complete hero detail review
